+++ src/commands/index.ts
@@ -144,6 +144,9 @@ import { verticalSpeedPriority } from './a32nx/verticalSpeedPriority';
 import { flyPadOS } from './a32nx/flyPadOS';
 import { simulationRate } from './support/simulationRate';
 import { dfd } from './general/dfd';
+import { direct } from './a32nx/direct';
+import { autoland } from './a32nx/autoland';
+import { holds } from './a32nx/holds';
 import { mcdu } from './a32nx/mcdu';
 
 const commands: CommandDefinition[] = [
@@ -291,6 +294,9 @@ const commands: CommandDefinition[] = [
     flyPadOS,
     simulationRate,
     dfd,
+    direct,
+    autoland,
+    holds,
     mcdu,
 ];