+++ .github/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 Update <small>_ September 2022</small>
 
-- feat: mcdu command for the detailed MCDU documentation (21/09/2022)
-- refactor: mcdu alias removed from remotemcdu command (21/09/2022)
+- feat: mcdu command for the detailed MCDU documentation (26/09/2022)
+- refactor: mcdu alias removed from remotemcdu command (26/09/2022)
+- feat: holds command for Holds documentation (26/09/2022)
+- feat: autoland command for Autoland documentation (26/09/2022)
+- feat: direct command for Direct documentation (26/09/2022)
 - fix: Changed image for assistance command (19/09/2022)
 - feat: add .dfd command (19/09/2022)
 - feat: executor added to the msgDelete log (07/09/2022)